Move globalstyles
